added visual Watchlist view
